Add auth to supermarket private controller

+++ src/main/java/com/heymart/supermarket/controller/PrivateSupermarketController.java
@@ -1,6 +1,8 @@
 package com.heymart.supermarket.controller;
 
+import com.heymart.supermarket.enums.SupermarketActions;
 import com.heymart.supermarket.model.Supermarket;
+import com.heymart.supermarket.service.AuthServiceClient;
 import com.heymart.supermarket.service.SupermarketService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,15 +15,22 @@ import java.util.List;
 @RequestMapping("/api/admin/supermarket")
 public class PrivateSupermarketController {
     private final SupermarketService service;
+    private final AuthServiceClient authServiceClient;
 
     @Autowired
-    public PrivateSupermarketController(SupermarketService supermarketService) {
-        this.service = supermarketService;
+    public PrivateSupermarketController(SupermarketService supermarketService, AuthServiceClient authServiceClient) {
+        this.service = supermarketService;this.authServiceClient = authServiceClient;
+
+
     }
 
     @PostMapping("/create")
-    public ResponseEntity<Object> create(@RequestBody Supermarket supermarket) {
-        try {
+    public ResponseEntity<Object> create(@RequestBody Supermarket supermarket, @RequestHeader(value = "Authorization", required = false) String authorizationHeader) {
+        if (!authServiceClient.verifyUserAuthorization(SupermarketActions.AUTH_CREATE_MANAGER.getValue(), authorizationHeader)) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
+            try {
             Supermarket newSupermarket = service.create(supermarket);
             return new ResponseEntity<>(newSupermarket, HttpStatus.OK);
         } catch (Exception e) {
@@ -30,7 +39,11 @@ public class PrivateSupermarketController {
     }
 
     @PostMapping("/update/{id}")
-    public ResponseEntity<Object> create(@PathVariable Long id, @RequestBody Supermarket supermarket) {
+    public ResponseEntity<Object> create(@PathVariable Long id, @RequestBody Supermarket supermarket, @RequestHeader(value = "Authorization", required = false) String authorizationHeader) {
+        if (!authServiceClient.verifyUserAuthorization(SupermarketActions.AUTH_CREATE_MANAGER.getValue(), authorizationHeader)) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             Supermarket updated = service.update(id, supermarket);
             return new ResponseEntity<>(updated, HttpStatus.OK);
@@ -40,7 +53,11 @@ public class PrivateSupermarketController {
     }
 
     @PostMapping("/delete/{id}")
-    public ResponseEntity<Object> delete(@PathVariable Long id) {
+    public ResponseEntity<Object> delete(@PathVariable Long id, @RequestHeader(value = "Authorization", required = false) String authorizationHeader) {
+        if (!authServiceClient.verifyUserAuthorization(SupermarketActions.AUTH_CREATE_MANAGER.getValue(), authorizationHeader)) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             service.delete(id);
             return new ResponseEntity<>("DELETED", HttpStatus.OK);
@@ -50,7 +67,11 @@ public class PrivateSupermarketController {
     }
 
     @GetMapping("/get/{id}")
-    public ResponseEntity<Object> getById(@PathVariable Long id) {
+    public ResponseEntity<Object> getById(@PathVariable Long id, @RequestHeader(value = "Authorization", required = false) String authorizationHeader) {
+        if (!authServiceClient.verifyUserAuthorization(SupermarketActions.AUTH_CREATE_MANAGER.getValue(), authorizationHeader)) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             Supermarket supermarket = service.getSupermarketFull(id);
             return new ResponseEntity<>(supermarket, HttpStatus.OK);
@@ -60,7 +81,11 @@ public class PrivateSupermarketController {
     }
 
     @GetMapping("/getAll")
-    public ResponseEntity<Object> getAll() {
+    public ResponseEntity<Object> getAll(@RequestHeader(value = "Authorization", required = false) String authorizationHeader) {
+        if (!authServiceClient.verifyUserAuthorization(SupermarketActions.AUTH_CREATE_MANAGER.getValue(), authorizationHeader)) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         List<Supermarket> supermarket = service.getAllSupermarketFull();
         return new ResponseEntity<>(supermarket, HttpStatus.OK);
     }
